refactor: Extract configuration management into dedicated module
This commit moves configuration-related functions and structs from main.rs into a new configuration module. The changes include:
- Creating a new `configuration/mod.rs` file
- Moving Args struct, config loading, and coinbase output processing functions
- Simplifying main.rs by importing configuration functions
- Maintaining existing configuration logic and error handling
